@@ -3,6 +3,10 @@
   "version": "0.1.0",
   "description": "A pseudonym registration Molgenis app",
   "private": false,
+  "authors": [
+    "Max Postema <https://github.com/MaxPostema>",
+    "Joris de Keijser <https://github.com/chinook25>"
+  ],
   "license": "LGPL-3.0",
   "scripts": {
     "serve": "vue-cli-service serve",
@@ -11,33 +15,30 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "core-js": "^3.6.5",
-    "vue": "^2.6.11"
+    "core-js": "^3.19.0",
+    "vue": "^3.0.0"
   },
   "devDependencies": {
-    "@types/jest": "^24.0.19",
-    "@typescript-eslint/eslint-plugin": "^4.18.0",
-    "@typescript-eslint/parser": "^4.18.0",
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-typescript": "~4.5.0",
-    "@vue/cli-plugin-unit-jest": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
+    "@types/jest": "^27.0.2",
+    "@typescript-eslint/eslint-plugin": "^5.2.0",
+    "@typescript-eslint/parser": "^5.2.0",
+    "@vue/cli-plugin-babel": "~4.5.14",
+    "@vue/cli-plugin-eslint": "~4.5.14",
+    "@vue/cli-plugin-typescript": "~4.5.14",
+    "@vue/cli-plugin-unit-jest": "~4.5.14",
+    "@vue/cli-service": "~4.5.14",
+    "@vue/compiler-sfc": "^3.0.0",
     "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/eslint-config-typescript": "^7.0.0",
-    "@vue/test-utils": "^1.0.3",
-    "eslint": "^6.7.2",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-vue": "^6.2.2",
+    "@vue/test-utils": "^2.0.0-0",
+    "eslint": "^7.0.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-vue": "^7.20.0",
     "generate-json-webpack-plugin": "^1.0.0",
-    "prettier": "^2.2.1",
-    "typescript": "~4.1.5",
-    "vue-eslint-parser": "^8.0.0",
-    "vue-template-compiler": "^2.6.11",
+    "jshint": "^2.13.1",
+    "prettier": "^2.4.1",
+    "typescript": "~4.4.4",
+    "vue-jest": "^5.0.0-0",
     "zip-webpack-plugin": "^4.0.1"
-  },
-  "authors": [
-    "Max Postema <https://github.com/MaxPostema>",
-    "Joris de Keijser <https://github.com/chinook25>"
-  ]
+  }
 }
