@@ -1,15 +1,12 @@
 <template>
-  <div id="app">
-    <img alt="Vue logo" src="./assets/logo.png" />
-    <HelloWorld msg="Welcome to Your Vue.js + TypeScript App" />
-  </div>
+  <HelloWorld msg="Welcome to Your Vue.js + TypeScript App" />
 </template>
 
 <script lang="ts">
-import Vue from 'vue';
+import {defineComponent} from 'vue';
 import HelloWorld from './components/HelloWorld.vue';
 
-export default Vue.extend({
+export default defineComponent({
   name: 'App',
   components: {
     HelloWorld
